Define form element aliases (e.g. 'integer' maps to 'numeric')

+++ demo_form.php
@@ -59,7 +59,7 @@ $form_def=array(
       ),
       'age'	=>array(
         'name'	=>"Age",
-	'type'	=>"numeric",
+	'type'	=>"integer",
       ),
     ),
   ),
